# Fix lock contention in FluidValue.Create