--- a/Fluid/Values/FluidValue.cs
+++ b/Fluid/Values/FluidValue.cs
@@ -6,6 +6,7 @@ using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Text.Encodings.Web;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Fluid.Values
@@ -14,9 +15,7 @@ namespace Fluid.Values
     {
         public abstract void WriteTo(TextWriter writer, TextEncoder encoder, CultureInfo cultureInfo);
 
-        public static Dictionary<Type, Type> _genericDictionaryTypeCache = new();
-        public static object _synLock = new();
-
+        private static Dictionary<Type, Type> _genericDictionaryTypeCache = new();
 
         [Conditional("DEBUG")]
         protected static void AssertWriteToParameters(TextWriter writer, TextEncoder encoder, CultureInfo cultureInfo)
@@ -182,30 +181,24 @@ namespace Fluid.Values
                     }
 
                     // Check if it's a more specific IDictionary<string, V>, e.g. JObject
-
-                    if (!_genericDictionaryTypeCache.TryGetValue(typeOfValue, out var genericType))
+                    var cache = _genericDictionaryTypeCache;
+                    if (!cache.TryGetValue(typeOfValue, out var genericType))
                     {
-                        lock (_synLock)
+                        foreach (var i in typeOfValue.GetInterfaces())
                         {
-                            if (!_genericDictionaryTypeCache.TryGetValue(typeOfValue, out genericType))
+                            if (i.IsGenericType && i.GetGenericTypeDefinition() == typeof(IDictionary<,>) && i.GetGenericArguments()[0] == typeof(string))
                             {
-                                foreach (var i in typeOfValue.GetInterfaces())
-                                {
-                                    if (i.IsGenericType && i.GetGenericTypeDefinition() == typeof(IDictionary<,>) && i.GetGenericArguments()[0] == typeof(string))
-                                    {
-                                        genericType = typeof(ObjectDictionaryFluidIndexable<>).MakeGenericType(i.GetGenericArguments()[1]);
-
-                                        // Swap the previous cache with a new copy to prevent locking on reads
-
-                                        var dictionary = new Dictionary<Type, Type>(_genericDictionaryTypeCache);
-                                        dictionary[typeOfValue] = genericType;
-                                        _genericDictionaryTypeCache = dictionary;
-
-                                        break;
-                                    }
-                                }
+                                genericType = typeof(ObjectDictionaryFluidIndexable<>).MakeGenericType(i.GetGenericArguments()[1]);
+                                break;
                             }
                         }
+
+                        // Swap the previous cache with a new copy to prevent locking on reads
+                        // make sure we also store null value so we don't try again
+                        Interlocked.CompareExchange(ref _genericDictionaryTypeCache, new Dictionary<Type, Type>(cache)
+                        {
+                            [typeOfValue] = genericType
+                        }, cache);
                     }
 
                     if (genericType != null)
@@ -222,11 +215,12 @@ namespace Fluid.Values
                             return new ArrayValue(enumerable);
 
                         case IList list:
-                            var values = new List<FluidValue>(list.Count);
-                            foreach (var item in list)
+                            var values = new FluidValue[list.Count];
+                            for (var i = 0; i < values.Length; i++)
                             {
-                                values.Add(Create(item, options));
+                                values[i] = Create(list[i], options);
                             }
+
                             return new ArrayValue(values);
 
                         case IEnumerable enumerable: