While doing performance checkup in Orchard (load test profiling agency theme) I noticed that there was quite big lock contention problem in Fluid, it originates from https://github.com/sebastienros/fluid/pull/387 where the logic never assigns anything on cache miss thus making the locking and looping happen every time for types like `ContentItem` (non-dictionary type). 

![image](https://user-images.githubusercontent.com/171892/147853913-485aefc7-f45a-4734-a144-376a61a4e351.png)

Fixed issue by using `Interlocked.CompareExchange`  (no locks needed) and also storing null as result for given type.  The dictionary was public, don't know why (Orchard doesn't seem to need it), made it private.

## Main

|         Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| CreateFromList | 287.9 ns | 1.22 ns | 1.08 ns | 0.0110 |     - |     - |     184 B |

## This PR

|         Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| CreateFromList | 46.98 ns | 0.979 ns | 1.048 ns | 0.0029 |     - |     - |      48 B |