@sebastienros 2022-01-01
There is a risk with the swap technique that multiple threads can never get to an agreement with all types in the dictionary. Example with two types, `A` and `B`.
Thread 1 checks for `A`, Thread 2 checks for `B`, none is there, T1 adds `A` replaces dictionary, T2 adds `B`, replaces dictionary, `A` is not there anymore, and this could go on. This is theory, but my previous lock was preventing this. Thread 2 would be blocked until Thread 1 was done.

So ideally the only thing to fix would be the `null` in the dictionary. In terms of perf no lock would be hit once the cache is warm.