@lahma 2022-01-01
https://docs.microsoft.com/en-us/dotnet/api/system.threading.interlocked.compareexchange?view=net-6.0#System_Threading_Interlocked_CompareExchange_System_Object__System_Object_System_Object_ - exchange only happens if field has same reference as in start, if not, someone did successful update and the current race loser will try again. But can also be changed back to double-checked locking.