@sebastienros 2022-01-02
I understand now, thanks for the explanation. There might be other places I could use that pattern then, I have used the "switch" technique in some places knowing the limitation I was explaining. This technique solves this quite nicely.